return all available tags in `/status.json`
Previously, tags that were not assigned to transcripts were not
available in the JSON because `.Site.Taxonomies.tags` returns
only tags that have been assigned to content/transcript.

To solve this, we create `all_tags.json` as part of the prebuild
step in order to have access to all the tags available from Optech.

This commit also switches the temporay source  from which we
were fetching the tags/topics as now we can get them directly
from "https://bitcoinops.org/topics.json"